requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
transformers==4.40.0
torch==2.3.0
langdetect==1.0.9
rank-bm25==0.2.2
pdfplumber==0.11.0
numpy==1.26.4
scikit-learn==1.4.2
pydantic==2.7.1
python-multipart==0.0.9
openai>=1.30.0
python-dotenv>=1.0.1
deep-translator==1.11.4
python-telegram-bot==20.7
httpx>=0.25.0
